uint8_t -> uint16_t, skip the first 7 BXs

+++ SimCalorimetry/HGCalSimProducers/interface/HGCDigitizerBase.cuh
@@ -5,8 +5,8 @@
 #include <curand.h>
 
 __global__
-void addNoise(int n, float* cellCharge, float* cellToa, bool weightMode, float* devRand, uint8_t* cellType, uint* word);
+void addNoise(int n, float* cellCharge, float* cellToa, bool weightMode, float* devRand, uint16_t* cellType, uint* word);
 
-void addNoiseWrapper(int n, float* cellCharge, float* cellToa, bool weightMode, float* devRand, uint8_t* cellType, uint* word,curandGenerator_t &d_gen);
+void addNoiseWrapper(int n, float* cellCharge, float* cellToa, bool weightMode, float* devRand, uint16_t* cellType, uint* word,curandGenerator_t &d_gen);
 
 #endif
